Harden detached Job replay and state isolation
